Save photo names in db and upload after submit is pressed
