+++ frontend/package.json
@@ -24,12 +24,12 @@
     "react-native-reanimated": "^1.2.0",
     "react-native-simple-radio-button": "^2.7.4",
     "react-native-table-component": "^1.2.1",
-    "react-native-uuid-generator": "^6.1.1",
     "react-native-vector-icons": "^6.6.0",
     "react-native-web": "^0.11.7",
     "react-navigation": "^4.0.10",
     "react-navigation-stack": "^1.9.3",
-    "react-navigation-tabs": "^2.5.5"
+    "react-navigation-tabs": "^2.5.5",
+    "uuid": "^3.4.0"
   },
   "devDependencies": {
     "@types/jest": "^24.9.1",
